{{ cookiecutter.project_name }}/pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "{{ cookiecutter.project_name.lower().replace('-', '_') }}"
version = "{{ cookiecutter.version }}"
description = "{{ cookiecutter.description }}"
authors = ["{{ cookiecutter.author_name }} <{{ cookiecutter.author_email }}>"]
packages = [
    { include = "{{ cookiecutter.project_name.lower().replace('-', '_') }}", from = "src" },
]

[tool.poetry.dependencies]
aiohttp-micro = {git = "https://github.com/clayman083/aiohttp-micro.git", rev = "v0.6.0"}
python = "^3.10"

[tool.poetry.dev-dependencies]
black = "19.10b0"
coverage = "5.5"
faker = "8.8.2"
flake8 = "3.9.2"
flake8-aaa = "0.12.0"
flake8-bugbear = "21.4.3"
flake8-builtins-unleashed = "1.3.1"
flake8-comprehensions = "3.5.0"
flake8-import-order = "0.18.1"
flake8-print = "4.0.0"
ipdb = "0.13.9"
ipython = "7.25.0"
marshmallow-jsonschema = "0.12.0"
mypy = "0.910"
pytest = "6.2.4"
pytest-aiohttp = "0.3.0"
pytest-cov = "2.12.1"
pytest-faker = "2.0.0"
pytest-fixture-maker = "0.2.0"
pytest-freezegun = "0.4.2"
pytest-mock = "3.6.1"
pytest-postgres = "0.7.0"
pytest-xdist = {extras = ["psutil"], version = "2.3.0"}
rope = "0.19.0"
types-click = "7.1.2"
